Update milestone check after "Part of a WIP Feature" label change
The `Part of a WIP Feature` label was changed to `status: do not merge` in the `woocommerce-android` and `woocommerce-ios` repos.

Since the Peril rule is used for other repos which may still have the `Part of a WIP Feature`, I tried making it so the update will still work if either label is used. I am actually not sure if I did it right or if I should simply have copied the entire block to run it again except with the separate constant and label name. Does that make sense?

<sup>(internal reference: paaHJt-244-p2)</sup>